feat: add client decoder and fromsqltext
