+++ src/error.rs
@@ -303,8 +303,14 @@ pub enum PgWireClientError {
     #[error("Error received from remote server: {0}")]
     RemoteError(Box<ErrorInfo>),
 
-    #[error("Error parse command tag")]
+    #[error("Error parse command tag: {0}")]
     InvalidTag(Box<dyn std::error::Error>),
+
+    #[error("Failed to parse data: {0}")]
+    FromSqlError(Box<dyn std::error::Error + Send + Sync>),
+
+    #[error("Index out of bounds")]
+    DataRowIndexOutOfBounds,
 }
 
 #[cfg(feature = "client-api")]
